2월 26일 Board 백엔드 수업

+++ Board/board/src/main/java/com/aloha/board/mapper/BaseMapper.java
@@ -4,11 +4,11 @@ import java.util.List;
 
 public interface BaseMapper<E> {
     List<E> list();
-    E select(int no);
+    E select(Long no);
     E selectById(String id);
     int insert(E boards);
     int update(E boards);
     int updateById(E boards);
-    int delete(int no);
+    int delete(Long no);
     int deleteById(String id);
 }
