+++ Board/board/src/main/java/com/aloha/board/service/BaseService.java
@@ -7,11 +7,11 @@ import com.github.pagehelper.PageInfo;
 public interface BaseService<E> {
     List<E> list();
     PageInfo<E> page(int page, int size);
-    E select(int no);
+    E select(Long no);
     E selectById(String id);
     boolean insert(E entity);
     boolean update(E entity);
     boolean updateById(E entity);
-    boolean delete(int no);
+    boolean delete(Long no);
     boolean deleteById(String id);
 }
